% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{measure}
\alias{measure}
\title{measure data set}
\usage{
data(measure)
}
\description{
This is a relatively small data set of anthropometric measurements. It contains
measurments for chest, waist and hip of 20 individuals, 10 males and 10 females. The
following variables are represented in columns:
\describe{
\item{chest}{chest circumference (distance around the waist) in inches}
\item{waist}{waist circumference in inches}
\item{hip}{hip circumference in inches}
\item{gender}{a factor: male or female}
}
}
\keyword{datasets}

